리팩터링
사용하지 않는 모듈 삭제
PEP8 준수하도록 코드 수정
visual.py 파일 모듈화
