+++ charts/__init__.py
@@ -1,0 +1,1 @@
+from .visual import show_smoking
